DEV: moving creation of wasm modult to enigma runtime
