+++ enigma-core/enclave/src/lib.rs
@@ -36,9 +36,6 @@ extern crate sputnikvm;
 extern crate sputnikvm_network_classic;
 extern crate wasmi;
 extern crate rustc_hex;
-/// This module builds Wasm code for contract deployment from the Wasm contract.
-/// The contract should be written in rust and then compiled to Wasm with wasm32-unknown-unknown target.
-/// The code is based on Parity wasm_utils::cli.
 extern crate pwasm_utils as wasm_utils;
 
 mod evm_t;
