+++ enigma-runtime-t/Cargo.toml
@@ -13,7 +13,8 @@ json-patch = { git = "https://github.com/enigmampc/json-patch.git", rev = "0.2.5
 serde_json = { git = "https://github.com/enigmampc/serde-json-sgx.git", rev = "1.0.39-sgx-1.0.8" }
 serde = { git = "https://github.com/mesalock-linux/serde-sgx.git", rev = "407aaaa2558dda5c2367d57ba40324c52cf9acc2", default-features = false, features=["serde_derive"] }
 wasmi = { git = "https://github.com/enigmampc/wasmi", rev = "0.4.2-sgx-1.0.8" }
-
+pwasm-utils = { git = "https://github.com/enigmampc/wasm-utils.git", rev = "0.5.0-sgx-1.0.8", default-features = false }
+parity-wasm = { git = "https://github.com/enigmampc/parity-wasm.git", branch = "enigma", default-features = false }
 
 sgx_tstd = { git = "https://github.com/baidu/rust-sgx-sdk.git", rev = "v1.0.8" }
 sgx_types = { git = "https://github.com/baidu/rust-sgx-sdk.git", rev = "v1.0.8" }
